backend/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-appconf==1.0.5
django-cors-headers==3.13.0
django-imagekit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pilkit==2.0
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
